+++ src/pages/minjiJeong/Main/Comment/Comment.js
@@ -1,44 +1,84 @@
 import React from 'react';
+import CommentList from './CommentList';
 import './Comment.scss';
 
 class Comment extends React.Component {
   constructor() {
     super();
     this.state = {
-      isLiked: false,
+      userId: '_minji.jeong',
+      commentList: [],
+      content: '',
     };
   }
 
-  toggleLike = () => {
+  getComment = e => {
     this.setState({
-      isLiked: !this.state.isLiked,
+      content: e.target.value,
+    });
+  };
+
+  uploadComment = e => {
+    this.state.commentList.push({
+      text: this.state.content,
+    });
+    this.setState({ content: '' });
+  };
+
+  handleEnter = e => {
+    if (e.code === 'Enter') {
+      this.uploadComment();
+    }
+  };
+
+  removeComment = deleteList => {
+    const newArr = this.state.commentList.filter(
+      (item, index) => index !== deleteList.props.listId
+    );
+
+    this.setState({
+      commentList: newArr,
     });
   };
 
   render() {
     return (
-      <div className="comments_row">
-        <p>
-          <span className="user_id">{this.props.userId}</span>
-          {this.props.content}
-        </p>
-        <div className="comments_menu">
-          <i
-            className={
-              this.state.isLiked
-                ? 'comment_like fas fa-heart'
-                : 'comment_like far fa-heart'
-            }
-            onClick={this.toggleLike}
-          ></i>
-          <i
-            className="comment_delete far fa-trash-alt"
-            onClick={() => {
-              this.props.removeComment(this);
-            }}
-          ></i>
+      <>
+        <div className="article_comments">
+          {/* 추가된 댓글 위치 */}
+          {this.state.commentList.map((el, index) => {
+            return (
+              <CommentList
+                key={index}
+                listId={index}
+                userId={this.state.userId}
+                content={el.text}
+                removeComment={this.removeComment}
+              />
+            );
+          })}
         </div>
-      </div>
+        <div className="time_ago">
+          <span className="pass_time">42</span>분 전
+        </div>
+        <div className="add_comment">
+          <textarea
+            cols="30"
+            rows="1"
+            placeholder="댓글 달기..."
+            onChange={this.getComment}
+            onKeyPress={this.handleEnter}
+            value={this.state.content}
+          ></textarea>
+          <button
+            className="btn_upload"
+            type="button"
+            onClick={this.uploadComment}
+          >
+            게시
+          </button>
+        </div>
+      </>
     );
   }
 }
