+++ src/pages/minjiJeong/Main/Comment/CommentList.scss
@@ -1,0 +1,41 @@
+@import '../../../../styles/variable.scss';
+
+.comments_row {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-direction: row;
+  width: 100%;
+  margin: 3px 0;
+}
+
+.comments_menu {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: row;
+
+  i {
+    &:hover {
+      cursor: pointer;
+    }
+  }
+
+  .fas {
+    color: #ed4956;
+  }
+
+  .far {
+    color: $theme-color-border;
+  }
+}
+
+.comment_like {
+  margin-right: 0;
+  transition: all ease 0.3s;
+}
+
+.comment_delete {
+  margin-left: 8px;
+  color: $theme-color-border;
+}
