+++ src/pages/minjiJeong/Main/Main.scss
@@ -219,41 +219,6 @@ nav {
   }
 }
 
-/*댓글*/
-.article_comments {
-  @include flex-center;
-  flex-direction: column;
-  padding: 0 15px;
-  font-size: 14px;
-}
-
-.time_ago {
-  margin-bottom: 10px;
-  padding: 4px 15px;
-  font-size: 13px;
-  font-weight: 600;
-  color: rgba(0, 0, 0, 0.3);
-}
-
-.add_comment {
-  display: flex;
-  flex-direction: row;
-  border-top: 1px solid $theme-color-border;
-
-  textarea {
-    width: 88%;
-    padding: 16px 10px;
-    font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
-  }
-}
-
-.btn_upload {
-  flex: 1;
-  font-weight: bold;
-  opacity: 0.4;
-  color: $theme-color-button;
-}
-
 /* Side Contents */
 .side_area {
   width: 300px;
